Begin multi-platform support
For now, the only script that works is mac. Linux /should/ work as
well, but I haven’t had time to test it. Windows definitely does not
work. I’ve included the dd for windows binary but In the future the
binary will be downloaded to the users computer from the internet in
the script. I also made a new central USBMaker script which is neat.