Remove 'dd.exe'
To ensure no software is redistributed, I removed the dd.exe from the
repo and downloaded & extracted it within the script directly from
chrysocome’s website. Also, a small, hidden file is created to check if
setup has been done yet. This will probably be replaced by just
checking for /MAC, /WINDOWS and /LINUX directories but I figured it’s a
good example on how to create files with python.